@@ -1,7 +1,10 @@
 """
-This module converts between AIMessage output formats for the Responses API.
+This module converts between AIMessage output formats, which are governed by the
+``output_version`` attribute on ChatOpenAI. Supported values are ``"v0"`` and
+``"responses/v1"``.
 
-ChatOpenAI v0.3 stores reasoning and tool outputs in AIMessage.additional_kwargs:
+``"v0"`` corresponds to the format as of ChatOpenAI v0.3. For the Responses API, it
+stores reasoning and tool outputs in AIMessage.additional_kwargs:
 
 .. code-block:: python
 
@@ -28,8 +31,9 @@ ChatOpenAI v0.3 stores reasoning and tool outputs in AIMessage.additional_kwargs
         id="msg_123",
     )
 
-To retain information about response item sequencing (and to accommodate multiple
-reasoning items), ChatOpenAI now stores these items in the content sequence:
+``"responses/v1"`` is only applicable to the Responses API. It retains information
+about response item sequencing and accommodates multiple reasoning items by
+representing these items in the content sequence:
 
 .. code-block:: python
 
@@ -57,18 +61,20 @@ There are other, small improvements as well-- e.g., we store message IDs on text
 content blocks, rather than on the AIMessage.id, which now stores the response ID.
 
 For backwards compatibility, this module provides functions to convert between the
-old and new formats. The functions are used internally by ChatOpenAI.
-
+formats. The functions are used internally by ChatOpenAI.
 """  # noqa: E501
 
 import json
-from typing import Union
+from collections.abc import Iterable, Iterator
+from typing import Any, Literal, Union, cast
 
-from langchain_core.messages import AIMessage
+from langchain_core.messages import AIMessage, is_data_content_block
+from langchain_core.messages import content as types
 
 _FUNCTION_CALL_IDS_MAP_KEY = "__openai_function_call_ids__"
 
 
+# v0.3 / Responses
 def _convert_to_v03_ai_message(
     message: AIMessage, has_reasoning: bool = False
 ) -> AIMessage:
@@ -141,115 +147,239 @@ def _convert_to_v03_ai_message(
     return message
 
 
-def _convert_from_v03_ai_message(message: AIMessage) -> AIMessage:
-    """Convert an old-style v0.3 AIMessage into the new content-block format."""
-    # Only update ChatOpenAI v0.3 AIMessages
-    # TODO: structure provenance into AIMessage
-    is_chatopenai_v03 = (
-        isinstance(message.content, list)
-        and all(isinstance(b, dict) for b in message.content)
-    ) and (
-        any(
-            item in message.additional_kwargs
-            for item in [
-                "reasoning",
-                "tool_outputs",
-                "refusal",
-                _FUNCTION_CALL_IDS_MAP_KEY,
-            ]
-        )
-        or (
-            isinstance(message.id, str)
-            and message.id.startswith("msg_")
-            and (response_id := message.response_metadata.get("id"))
-            and isinstance(response_id, str)
-            and response_id.startswith("resp_")
-        )
-    )
-    if not is_chatopenai_v03:
-        return message
+# v1 / Chat Completions
+def _convert_from_v1_to_chat_completions(message: AIMessage) -> AIMessage:
+    """Convert a v1 message to the Chat Completions format."""
+    if isinstance(message.content, list):
+        new_content: list = []
+        for block in message.content:
+            if isinstance(block, dict):
+                block_type = block.get("type")
+                if block_type == "text":
+                    # Strip annotations
+                    new_content.append({"type": "text", "text": block["text"]})
+                elif block_type in ("reasoning", "tool_call"):
+                    pass
+                else:
+                    new_content.append(block)
+            else:
+                new_content.append(block)
+        return message.model_copy(update={"content": new_content})
 
-    content_order = [
-        "reasoning",
-        "code_interpreter_call",
-        "mcp_call",
-        "image_generation_call",
-        "text",
-        "refusal",
-        "function_call",
-        "computer_call",
-        "mcp_list_tools",
-        "mcp_approval_request",
-        # N. B. "web_search_call" and "file_search_call" were not passed back in
-        # in v0.3
-    ]
+    return message
 
-    # Build a bucket for every known block type
-    buckets: dict[str, list] = {key: [] for key in content_order}
-    unknown_blocks = []
 
-    # Reasoning
-    if reasoning := message.additional_kwargs.get("reasoning"):
-        buckets["reasoning"].append(reasoning)
+# v1 / Responses
+def _convert_annotation_from_v1(annotation: types.Annotation) -> dict[str, Any]:
+    if annotation["type"] == "citation":
+        new_ann: dict[str, Any] = {}
+        for field in ("end_index", "start_index"):
+            if field in annotation:
+                new_ann[field] = annotation[field]
 
-    # Refusal
-    if refusal := message.additional_kwargs.get("refusal"):
-        buckets["refusal"].append({"type": "refusal", "refusal": refusal})
-
-    # Text
-    for block in message.content:
-        if isinstance(block, dict) and block.get("type") == "text":
-            block_copy = block.copy()
-            if isinstance(message.id, str) and message.id.startswith("msg_"):
-                block_copy["id"] = message.id
-            buckets["text"].append(block_copy)
+        if "url" in annotation:
+            # URL citation
+            if "title" in annotation:
+                new_ann["title"] = annotation["title"]
+            new_ann["type"] = "url_citation"
+            new_ann["url"] = annotation["url"]
         else:
-            unknown_blocks.append(block)
+            # Document citation
+            new_ann["type"] = "file_citation"
+            if "title" in annotation:
+                new_ann["filename"] = annotation["title"]
 
-    # Function calls
-    function_call_ids = message.additional_kwargs.get(_FUNCTION_CALL_IDS_MAP_KEY)
-    for tool_call in message.tool_calls:
-        function_call = {
-            "type": "function_call",
-            "name": tool_call["name"],
-            "arguments": json.dumps(tool_call["args"], ensure_ascii=False),
-            "call_id": tool_call["id"],
-        }
-        if function_call_ids is not None and (
-            _id := function_call_ids.get(tool_call["id"])
-        ):
-            function_call["id"] = _id
-        buckets["function_call"].append(function_call)
+        if extra_fields := annotation.get("extras"):
+            for field, value in extra_fields.items():
+                new_ann[field] = value
 
-    # Tool outputs
-    tool_outputs = message.additional_kwargs.get("tool_outputs", [])
-    for block in tool_outputs:
-        if isinstance(block, dict) and (key := block.get("type")) and key in buckets:
-            buckets[key].append(block)
-        else:
-            unknown_blocks.append(block)
+        return new_ann
 
-    # Re-assemble the content list in the canonical order
-    new_content = []
-    for key in content_order:
-        new_content.extend(buckets[key])
-    new_content.extend(unknown_blocks)
+    elif annotation["type"] == "non_standard_annotation":
+        return annotation["value"]
 
-    new_additional_kwargs = dict(message.additional_kwargs)
-    new_additional_kwargs.pop("reasoning", None)
-    new_additional_kwargs.pop("refusal", None)
-    new_additional_kwargs.pop("tool_outputs", None)
-
-    if "id" in message.response_metadata:
-        new_id = message.response_metadata["id"]
     else:
-        new_id = message.id
+        return dict(annotation)
 
-    return message.model_copy(
-        update={
-            "content": new_content,
-            "additional_kwargs": new_additional_kwargs,
-            "id": new_id,
-        },
-        deep=False,
+
+def _implode_reasoning_blocks(blocks: list[dict[str, Any]]) -> Iterable[dict[str, Any]]:
+    i = 0
+    n = len(blocks)
+
+    while i < n:
+        block = blocks[i]
+
+        # Skip non-reasoning blocks or blocks already in Responses format
+        if block.get("type") != "reasoning" or "summary" in block:
+            yield dict(block)
+            i += 1
+            continue
+        elif "reasoning" not in block and "summary" not in block:
+            # {"type": "reasoning", "id": "rs_..."}
+            oai_format = {**block, "summary": []}
+            if "extras" in oai_format:
+                oai_format.update(oai_format.pop("extras"))
+            oai_format["type"] = oai_format.pop("type", "reasoning")
+            if "encrypted_content" in oai_format:
+                oai_format["encrypted_content"] = oai_format.pop("encrypted_content")
+            yield oai_format
+            i += 1
+            continue
+        else:
+            pass
+
+        summary: list[dict[str, str]] = [
+            {"type": "summary_text", "text": block.get("reasoning", "")}
+        ]
+        # 'common' is every field except the exploded 'reasoning'
+        common = {k: v for k, v in block.items() if k != "reasoning"}
+        if "extras" in common:
+            common.update(common.pop("extras"))
+
+        i += 1
+        while i < n:
+            next_ = blocks[i]
+            if next_.get("type") == "reasoning" and "reasoning" in next_:
+                summary.append(
+                    {"type": "summary_text", "text": next_.get("reasoning", "")}
+                )
+                i += 1
+            else:
+                break
+
+        merged = dict(common)
+        merged["summary"] = summary
+        merged["type"] = merged.pop("type", "reasoning")
+        yield merged
+
+
+def _consolidate_calls(
+    items: Iterable[dict[str, Any]],
+    call_name: Literal["web_search_call", "code_interpreter_call"],
+    result_name: Literal["web_search_result", "code_interpreter_result"],
+) -> Iterator[dict[str, Any]]:
+    """
+    Generator that walks through *items* and, whenever it meets the pair
+
+        {"type": "web_search_call",    "id": X, ...}
+        {"type": "web_search_result",  "id": X}
+
+    merges them into
+
+        {"id": X,
+         "action": …,
+         "status": …,
+         "type": "web_search_call"}
+
+    keeping every other element untouched.
+    """
+    items = iter(items)  # make sure we have a true iterator
+    for current in items:
+        # Only a call can start a pair worth collapsing
+        if current.get("type") != call_name:
+            yield current
+            continue
+
+        try:
+            nxt = next(items)  # look-ahead one element
+        except StopIteration:  # no “result” – just yield the call back
+            yield current
+            break
+
+        # If this really is the matching “result” – collapse
+        if nxt.get("type") == result_name and nxt.get("id") == current.get("id"):
+            if call_name == "web_search_call":
+                collapsed = {"id": current["id"]}
+                if "action" in current:
+                    collapsed["action"] = current["action"]
+                collapsed["status"] = current["status"]
+                collapsed["type"] = "web_search_call"
+
+            if call_name == "code_interpreter_call":
+                collapsed = {"id": current["id"]}
+                for key in ("code", "container_id"):
+                    if key in current:
+                        collapsed[key] = current[key]
+                    elif key in current.get("extras", {}):
+                        collapsed[key] = current["extras"][key]
+                    else:
+                        pass
+
+                for key in ("outputs", "status"):
+                    if key in nxt:
+                        collapsed[key] = nxt[key]
+                    elif key in nxt.get("extras", {}):
+                        collapsed[key] = nxt["extras"][key]
+                    else:
+                        pass
+                collapsed["type"] = "code_interpreter_call"
+
+            yield collapsed
+
+        else:
+            # Not a matching pair – emit both, in original order
+            yield current
+            yield nxt
+
+
+def _convert_from_v1_to_responses(
+    content: list[types.ContentBlock], tool_calls: list[types.ToolCall]
+) -> list[dict[str, Any]]:
+    new_content: list = []
+    for block in content:
+        if block["type"] == "text" and "annotations" in block:
+            # Need a copy because we’re changing the annotations list
+            new_block = dict(block)
+            new_block["annotations"] = [
+                _convert_annotation_from_v1(a) for a in block["annotations"]
+            ]
+            new_content.append(new_block)
+        elif block["type"] == "tool_call":
+            new_block = {"type": "function_call", "call_id": block["id"]}
+            if "extras" in block and "item_id" in block["extras"]:
+                new_block["id"] = block["extras"]["item_id"]
+            if "name" in block:
+                new_block["name"] = block["name"]
+            if "extras" in block and "arguments" in block["extras"]:
+                new_block["arguments"] = block["extras"]["arguments"]
+            if any(key not in block for key in ("name", "arguments")):
+                matching_tool_calls = [
+                    call for call in tool_calls if call["id"] == block["id"]
+                ]
+                if matching_tool_calls:
+                    tool_call = matching_tool_calls[0]
+                    if "name" not in block:
+                        new_block["name"] = tool_call["name"]
+                    if "arguments" not in block:
+                        new_block["arguments"] = json.dumps(tool_call["args"])
+            new_content.append(new_block)
+        elif (
+            is_data_content_block(cast(dict, block))
+            and block["type"] == "image"
+            and "base64" in block
+            and isinstance(block.get("id"), str)
+            and block["id"].startswith("ig_")
+        ):
+            new_block = {"type": "image_generation_call", "result": block["base64"]}
+            for extra_key in ("id", "status"):
+                if extra_key in block:
+                    new_block[extra_key] = block[extra_key]  # type: ignore[typeddict-item]
+                elif extra_key in block.get("extras", {}):
+                    new_block[extra_key] = block["extras"][extra_key]
+            new_content.append(new_block)
+        elif block["type"] == "non_standard" and "value" in block:
+            new_content.append(block["value"])
+        else:
+            new_content.append(block)
+
+    new_content = list(_implode_reasoning_blocks(new_content))
+    new_content = list(
+        _consolidate_calls(new_content, "web_search_call", "web_search_result")
     )
+    new_content = list(
+        _consolidate_calls(
+            new_content, "code_interpreter_call", "code_interpreter_result"
+        )
+    )
+
+    return new_content
