@@ -59,7 +59,7 @@ async def test_generic_fake_chat_model_stream() -> None:
     assert chunks == [
         _any_id_ai_message_chunk(content="hello"),
         _any_id_ai_message_chunk(content=" "),
-        _any_id_ai_message_chunk(content="goodbye"),
+        _any_id_ai_message_chunk(content="goodbye", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in chunks}) == 1
 
@@ -67,7 +67,7 @@ async def test_generic_fake_chat_model_stream() -> None:
     assert chunks == [
         _any_id_ai_message_chunk(content="hello"),
         _any_id_ai_message_chunk(content=" "),
-        _any_id_ai_message_chunk(content="goodbye"),
+        _any_id_ai_message_chunk(content="goodbye", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in chunks}) == 1
 
@@ -79,6 +79,7 @@ async def test_generic_fake_chat_model_stream() -> None:
     assert chunks == [
         _any_id_ai_message_chunk(content="", additional_kwargs={"foo": 42}),
         _any_id_ai_message_chunk(content="", additional_kwargs={"bar": 24}),
+        _any_id_ai_message_chunk(content="", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in chunks}) == 1
 
@@ -97,7 +98,8 @@ async def test_generic_fake_chat_model_stream() -> None:
 
     assert chunks == [
         _any_id_ai_message_chunk(
-            content="", additional_kwargs={"function_call": {"name": "move_file"}}
+            content="",
+            additional_kwargs={"function_call": {"name": "move_file"}},
         ),
         _any_id_ai_message_chunk(
             content="",
@@ -114,6 +116,7 @@ async def test_generic_fake_chat_model_stream() -> None:
                 "function_call": {"arguments": '\n  "destination_path": "bar"\n}'},
             },
         ),
+        _any_id_ai_message_chunk(content="", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in chunks}) == 1
 
@@ -134,6 +137,7 @@ async def test_generic_fake_chat_model_stream() -> None:
             }
         },
         id=chunks[0].id,
+        chunk_position="last",
     )
 
 
@@ -148,7 +152,7 @@ async def test_generic_fake_chat_model_astream_log() -> None:
     assert final.state["streamed_output"] == [
         _any_id_ai_message_chunk(content="hello"),
         _any_id_ai_message_chunk(content=" "),
-        _any_id_ai_message_chunk(content="goodbye"),
+        _any_id_ai_message_chunk(content="goodbye", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in final.state["streamed_output"]}) == 1
 
@@ -205,7 +209,7 @@ async def test_callback_handlers() -> None:
     assert results == [
         _any_id_ai_message_chunk(content="hello"),
         _any_id_ai_message_chunk(content=" "),
-        _any_id_ai_message_chunk(content="goodbye"),
+        _any_id_ai_message_chunk(content="goodbye", chunk_position="last"),
     ]
     assert tokens == ["hello", " ", "goodbye"]
     assert len({chunk.id for chunk in results}) == 1
