@@ -14,11 +14,15 @@ from langchain_core.language_models import (
     ParrotFakeChatModel,
 )
 from langchain_core.language_models._utils import _normalize_messages
-from langchain_core.language_models.fake_chat_models import FakeListChatModelError
+from langchain_core.language_models.fake_chat_models import (
+    FakeListChatModelError,
+    GenericFakeChatModel,
+)
 from langchain_core.messages import (
     AIMessage,
     AIMessageChunk,
     BaseMessage,
+    BaseMessageChunk,
     HumanMessage,
     SystemMessage,
 )
@@ -38,6 +42,37 @@ from tests.unit_tests.stubs import _any_id_ai_message, _any_id_ai_message_chunk
 
 if TYPE_CHECKING:
     from langchain_core.outputs.llm_result import LLMResult
+
+
+def _content_blocks_equal_ignore_id(
+    actual: Union[str, list[Any]], expected: Union[str, list[Any]]
+) -> bool:
+    """Compare content blocks, ignoring auto-generated `id` fields.
+
+    Args:
+        actual: Actual content from response (string or list of content blocks).
+        expected: Expected content to compare against (string or list of blocks).
+
+    Returns:
+        True if content matches (excluding `id` fields), False otherwise.
+
+    """
+    if isinstance(actual, str) or isinstance(expected, str):
+        return actual == expected
+
+    if len(actual) != len(expected):
+        return False
+    for actual_block, expected_block in zip(actual, expected):
+        actual_without_id = (
+            {k: v for k, v in actual_block.items() if k != "id"}
+            if isinstance(actual_block, dict) and "id" in actual_block
+            else actual_block
+        )
+
+        if actual_without_id != expected_block:
+            return False
+
+    return True
 
 
 @pytest.fixture
@@ -141,7 +176,7 @@ async def test_stream_error_callback() -> None:
 
 
 async def test_astream_fallback_to_ainvoke() -> None:
-    """Test astream uses appropriate implementation."""
+    """Test `astream()` uses appropriate implementation."""
 
     class ModelWithGenerate(BaseChatModel):
         @override
@@ -198,7 +233,9 @@ async def test_astream_implementation_fallback_to_stream() -> None:
         ) -> Iterator[ChatGenerationChunk]:
             """Stream the output of the model."""
             yield ChatGenerationChunk(message=AIMessageChunk(content="a"))
-            yield ChatGenerationChunk(message=AIMessageChunk(content="b"))
+            yield ChatGenerationChunk(
+                message=AIMessageChunk(content="b", chunk_position="last")
+            )
 
         @property
         def _llm_type(self) -> str:
@@ -207,15 +244,19 @@ async def test_astream_implementation_fallback_to_stream() -> None:
     model = ModelWithSyncStream()
     chunks = list(model.stream("anything"))
     assert chunks == [
-        _any_id_ai_message_chunk(content="a"),
-        _any_id_ai_message_chunk(content="b"),
+        _any_id_ai_message_chunk(
+            content="a",
+        ),
+        _any_id_ai_message_chunk(content="b", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in chunks}) == 1
     assert type(model)._astream == BaseChatModel._astream
     astream_chunks = [chunk async for chunk in model.astream("anything")]
     assert astream_chunks == [
-        _any_id_ai_message_chunk(content="a"),
-        _any_id_ai_message_chunk(content="b"),
+        _any_id_ai_message_chunk(
+            content="a",
+        ),
+        _any_id_ai_message_chunk(content="b", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in astream_chunks}) == 1
 
@@ -244,7 +285,9 @@ async def test_astream_implementation_uses_astream() -> None:
         ) -> AsyncIterator[ChatGenerationChunk]:
             """Stream the output of the model."""
             yield ChatGenerationChunk(message=AIMessageChunk(content="a"))
-            yield ChatGenerationChunk(message=AIMessageChunk(content="b"))
+            yield ChatGenerationChunk(
+                message=AIMessageChunk(content="b", chunk_position="last")
+            )
 
         @property
         def _llm_type(self) -> str:
@@ -253,8 +296,10 @@ async def test_astream_implementation_uses_astream() -> None:
     model = ModelWithAsyncStream()
     chunks = [chunk async for chunk in model.astream("anything")]
     assert chunks == [
-        _any_id_ai_message_chunk(content="a"),
-        _any_id_ai_message_chunk(content="b"),
+        _any_id_ai_message_chunk(
+            content="a",
+        ),
+        _any_id_ai_message_chunk(content="b", chunk_position="last"),
     ]
     assert len({chunk.id for chunk in chunks}) == 1
 
@@ -427,11 +472,12 @@ class FakeChatModelStartTracer(FakeTracer):
 
 
 def test_trace_images_in_openai_format() -> None:
-    """Test that images are traced in OpenAI format."""
+    """Test that images are traced in OpenAI Chat Completions format."""
     llm = ParrotFakeChatModel()
     messages = [
         {
             "role": "user",
+            # v0 format
             "content": [
                 {
                     "type": "image",
@@ -442,7 +488,7 @@ def test_trace_images_in_openai_format() -> None:
         }
     ]
     tracer = FakeChatModelStartTracer()
-    response = llm.invoke(messages, config={"callbacks": [tracer]})
+    llm.invoke(messages, config={"callbacks": [tracer]})
     assert tracer.messages == [
         [
             [
@@ -457,19 +503,51 @@ def test_trace_images_in_openai_format() -> None:
             ]
         ]
     ]
-    # Test no mutation
-    assert response.content == [
-        {
+
+
+def test_content_block_transformation_v0_to_v1_image() -> None:
+    """Test that v0 format image content blocks are transformed to v1 format."""
+    # Create a message with v0 format image content
+    image_message = AIMessage(
+        content=[
+            {
+                "type": "image",
+                "source_type": "url",
+                "url": "https://example.com/image.png",
+            }
+        ]
+    )
+
+    llm = GenericFakeChatModel(messages=iter([image_message]), output_version="v1")
+    response = llm.invoke("test")
+
+    # With v1 output_version, .content should be transformed
+    # Check structure, ignoring auto-generated IDs
+    assert len(response.content) == 1
+    content_block = response.content[0]
+    if isinstance(content_block, dict) and "id" in content_block:
+        # Remove auto-generated id for comparison
+        content_without_id = {k: v for k, v in content_block.items() if k != "id"}
+        expected_content = {
             "type": "image",
-            "source_type": "url",
             "url": "https://example.com/image.png",
         }
-    ]
+        assert content_without_id == expected_content
+    else:
+        assert content_block == {
+            "type": "image",
+            "url": "https://example.com/image.png",
+        }
 
 
-def test_trace_content_blocks_with_no_type_key() -> None:
-    """Test that we add a ``type`` key to certain content blocks that don't have one."""
-    llm = ParrotFakeChatModel()
+@pytest.mark.parametrize("output_version", ["v0", "v1"])
+def test_trace_content_blocks_with_no_type_key(output_version: str) -> None:
+    """Test behavior of content blocks that don't have a `type` key.
+
+    Only for blocks with one key, in which case, the name of the key is used as `type`.
+
+    """
+    llm = ParrotFakeChatModel(output_version=output_version)
     messages = [
         {
             "role": "user",
@@ -504,153 +582,378 @@ def test_trace_content_blocks_with_no_type_key() -> None:
             ]
         ]
     ]
-    # Test no mutation
-    assert response.content == [
+
+    if output_version == "v0":
+        assert response.content == [
+            {
+                "type": "text",
+                "text": "Hello",
+            },
+            {
+                "cachePoint": {"type": "default"},
+            },
+        ]
+    else:
+        assert response.content == [
+            {
+                "type": "text",
+                "text": "Hello",
+            },
+            {
+                "type": "non_standard",
+                "value": {
+                    "cachePoint": {"type": "default"},
+                },
+            },
+        ]
+
+    assert response.content_blocks == [
         {
             "type": "text",
             "text": "Hello",
         },
         {
-            "cachePoint": {"type": "default"},
+            "type": "non_standard",
+            "value": {
+                "cachePoint": {"type": "default"},
+            },
         },
     ]
 
 
 def test_extend_support_to_openai_multimodal_formats() -> None:
-    """Test that chat models normalize OpenAI file and audio inputs."""
-    llm = ParrotFakeChatModel()
-    messages = [
-        {
-            "role": "user",
-            "content": [
-                {"type": "text", "text": "Hello"},
-                {
-                    "type": "image_url",
-                    "image_url": {"url": "https://example.com/image.png"},
+    """Test normalizing OpenAI audio, image, and file inputs to v1."""
+    # Audio and file only (chat model default)
+    messages = HumanMessage(
+        content=[
+            {"type": "text", "text": "Hello"},
+            {  # audio-base64
+                "type": "input_audio",
+                "input_audio": {
+                    "format": "wav",
+                    "data": "<base64 string>",
                 },
-                {
-                    "type": "image_url",
-                    "image_url": {"url": "data:image/jpeg;base64,/9j/4AAQSkZJRg..."},
+            },
+            {  # file-base64
+                "type": "file",
+                "file": {
+                    "filename": "draconomicon.pdf",
+                    "file_data": "data:application/pdf;base64,<base64 string>",
                 },
-                {
-                    "type": "file",
-                    "file": {
-                        "filename": "draconomicon.pdf",
-                        "file_data": "data:application/pdf;base64,<base64 string>",
-                    },
-                },
-                {
-                    "type": "file",
-                    "file": {
-                        "file_data": "data:application/pdf;base64,<base64 string>",
-                    },
-                },
-                {
-                    "type": "file",
-                    "file": {"file_id": "<file id>"},
-                },
-                {
-                    "type": "input_audio",
-                    "input_audio": {"data": "<base64 data>", "format": "wav"},
-                },
-            ],
-        },
-    ]
-    expected_content = [
-        {"type": "text", "text": "Hello"},
-        {
-            "type": "image_url",
-            "image_url": {"url": "https://example.com/image.png"},
-        },
-        {
-            "type": "image_url",
-            "image_url": {"url": "data:image/jpeg;base64,/9j/4AAQSkZJRg..."},
-        },
-        {
-            "type": "file",
-            "source_type": "base64",
-            "data": "<base64 string>",
-            "mime_type": "application/pdf",
-            "filename": "draconomicon.pdf",
-        },
-        {
-            "type": "file",
-            "source_type": "base64",
-            "data": "<base64 string>",
-            "mime_type": "application/pdf",
-        },
-        {
-            "type": "file",
-            "file": {"file_id": "<file id>"},
-        },
-        {
-            "type": "audio",
-            "source_type": "base64",
-            "data": "<base64 data>",
-            "mime_type": "audio/wav",
-        },
-    ]
-    response = llm.invoke(messages)
-    assert response.content == expected_content
+            },
+            {  # file-id
+                "type": "file",
+                "file": {"file_id": "<file id>"},
+            },
+        ]
+    )
 
-    # Test no mutation
-    assert messages[0]["content"] == [
-        {"type": "text", "text": "Hello"},
-        {
-            "type": "image_url",
-            "image_url": {"url": "https://example.com/image.png"},
-        },
-        {
-            "type": "image_url",
-            "image_url": {"url": "data:image/jpeg;base64,/9j/4AAQSkZJRg..."},
-        },
-        {
-            "type": "file",
-            "file": {
-                "filename": "draconomicon.pdf",
-                "file_data": "data:application/pdf;base64,<base64 string>",
+    expected_content_messages = HumanMessage(
+        content=[
+            {"type": "text", "text": "Hello"},  # TextContentBlock
+            {  # AudioContentBlock
+                "type": "audio",
+                "base64": "<base64 string>",
+                "mime_type": "audio/wav",
             },
-        },
-        {
-            "type": "file",
-            "file": {
-                "file_data": "data:application/pdf;base64,<base64 string>",
+            {  # FileContentBlock
+                "type": "file",
+                "base64": "<base64 string>",
+                "mime_type": "application/pdf",
+                "extras": {"filename": "draconomicon.pdf"},
             },
-        },
-        {
-            "type": "file",
-            "file": {"file_id": "<file id>"},
-        },
-        {
-            "type": "input_audio",
-            "input_audio": {"data": "<base64 data>", "format": "wav"},
-        },
-    ]
+            {  # ...
+                "type": "file",
+                "file_id": "<file id>",
+            },
+        ]
+    )
+
+    normalized_content = _normalize_messages([messages])
+
+    # Check structure, ignoring auto-generated IDs
+    assert len(normalized_content) == 1
+    normalized_message = normalized_content[0]
+    assert len(normalized_message.content) == len(expected_content_messages.content)
+
+    assert _content_blocks_equal_ignore_id(
+        normalized_message.content, expected_content_messages.content
+    )
+
+    messages = HumanMessage(
+        content=[
+            {"type": "text", "text": "Hello"},
+            {  # image-url
+                "type": "image_url",
+                "image_url": {"url": "https://example.com/image.png"},
+            },
+            {  # image-base64
+                "type": "image_url",
+                "image_url": {"url": "data:image/jpeg;base64,/9j/4AAQSkZJRg..."},
+            },
+            {  # audio-base64
+                "type": "input_audio",
+                "input_audio": {
+                    "format": "wav",
+                    "data": "<base64 string>",
+                },
+            },
+            {  # file-base64
+                "type": "file",
+                "file": {
+                    "filename": "draconomicon.pdf",
+                    "file_data": "data:application/pdf;base64,<base64 string>",
+                },
+            },
+            {  # file-id
+                "type": "file",
+                "file": {"file_id": "<file id>"},
+            },
+        ]
+    )
+
+    expected_content_messages = HumanMessage(
+        content=[
+            {"type": "text", "text": "Hello"},  # TextContentBlock
+            {  # image-url passes through
+                "type": "image_url",
+                "image_url": {"url": "https://example.com/image.png"},
+            },
+            {  # image-url passes through with inline data
+                "type": "image_url",
+                "image_url": {"url": "data:image/jpeg;base64,/9j/4AAQSkZJRg..."},
+            },
+            {  # AudioContentBlock
+                "type": "audio",
+                "base64": "<base64 string>",
+                "mime_type": "audio/wav",
+            },
+            {  # FileContentBlock
+                "type": "file",
+                "base64": "<base64 string>",
+                "mime_type": "application/pdf",
+                "extras": {"filename": "draconomicon.pdf"},
+            },
+            {  # ...
+                "type": "file",
+                "file_id": "<file id>",
+            },
+        ]
+    )
+
+    normalized_content = _normalize_messages([messages])
+
+    # Check structure, ignoring auto-generated IDs
+    assert len(normalized_content) == 1
+    normalized_message = normalized_content[0]
+    assert len(normalized_message.content) == len(expected_content_messages.content)
+
+    assert _content_blocks_equal_ignore_id(
+        normalized_message.content, expected_content_messages.content
+    )
 
 
 def test_normalize_messages_edge_cases() -> None:
-    # Test some blocks that should pass through
+    # Test behavior of malformed/unrecognized content blocks
+
     messages = [
         HumanMessage(
             content=[
                 {
-                    "type": "file",
-                    "file": "uri",
+                    "type": "input_image",  # Responses API type; not handled
+                    "image_url": "uri",
                 },
                 {
-                    "type": "input_file",
+                    # Standard OpenAI Chat Completions type but malformed structure
+                    "type": "input_audio",
+                    "input_audio": "uri",  # Should be nested in `audio`
+                },
+                {
+                    "type": "file",
+                    "file": "uri",  # `file` should be a dict for Chat Completions
+                },
+                {
+                    "type": "input_file",  # Responses API type; not handled
                     "file_data": "uri",
                     "filename": "file-name",
-                },
-                {
-                    "type": "input_audio",
-                    "input_audio": "uri",
-                },
-                {
-                    "type": "input_image",
-                    "image_url": "uri",
                 },
             ]
         )
     ]
+
     assert messages == _normalize_messages(messages)
+
+
+def test_normalize_messages_v1_content_blocks_unchanged() -> None:
+    """Test passing v1 content blocks to `_normalize_messages()` leaves unchanged."""
+    input_messages = [
+        HumanMessage(
+            content=[
+                {
+                    "type": "text",
+                    "text": "Hello world",
+                },
+                {
+                    "type": "image",
+                    "url": "https://example.com/image.png",
+                    "mime_type": "image/png",
+                },
+                {
+                    "type": "audio",
+                    "base64": "base64encodedaudiodata",
+                    "mime_type": "audio/wav",
+                },
+                {
+                    "type": "file",
+                    "id": "file_123",
+                },
+                {
+                    "type": "reasoning",
+                    "reasoning": "Let me think about this...",
+                },
+            ]
+        )
+    ]
+
+    result = _normalize_messages(input_messages)
+
+    # Verify the result is identical to the input (message should not be copied)
+    assert len(result) == 1
+    assert result[0] is input_messages[0]
+    assert result[0].content == input_messages[0].content
+
+
+def test_output_version_invoke(monkeypatch: Any) -> None:
+    messages = [AIMessage("hello")]
+
+    llm = GenericFakeChatModel(messages=iter(messages), output_version="v1")
+    response = llm.invoke("hello")
+    assert response.content == [{"type": "text", "text": "hello"}]
+    assert response.response_metadata["output_version"] == "v1"
+
+    llm = GenericFakeChatModel(messages=iter(messages))
+    response = llm.invoke("hello")
+    assert response.content == "hello"
+
+    monkeypatch.setenv("LC_OUTPUT_VERSION", "v1")
+    llm = GenericFakeChatModel(messages=iter(messages))
+    response = llm.invoke("hello")
+    assert response.content == [{"type": "text", "text": "hello"}]
+    assert response.response_metadata["output_version"] == "v1"
+
+
+# -- v1 output version tests --
+
+
+async def test_output_version_ainvoke(monkeypatch: Any) -> None:
+    messages = [AIMessage("hello")]
+
+    # v0
+    llm = GenericFakeChatModel(messages=iter(messages))
+    response = await llm.ainvoke("hello")
+    assert response.content == "hello"
+
+    # v1
+    llm = GenericFakeChatModel(messages=iter(messages), output_version="v1")
+    response = await llm.ainvoke("hello")
+    assert response.content == [{"type": "text", "text": "hello"}]
+    assert response.response_metadata["output_version"] == "v1"
+
+    # v1 from env var
+    monkeypatch.setenv("LC_OUTPUT_VERSION", "v1")
+    llm = GenericFakeChatModel(messages=iter(messages))
+    response = await llm.ainvoke("hello")
+    assert response.content == [{"type": "text", "text": "hello"}]
+    assert response.response_metadata["output_version"] == "v1"
+
+
+def test_output_version_stream(monkeypatch: Any) -> None:
+    messages = [AIMessage("foo bar")]
+
+    # v0
+    llm = GenericFakeChatModel(messages=iter(messages))
+    full = None
+    for chunk in llm.stream("hello"):
+        assert isinstance(chunk, AIMessageChunk)
+        assert isinstance(chunk.content, str)
+        assert chunk.content
+        full = chunk if full is None else full + chunk
+    assert isinstance(full, AIMessageChunk)
+    assert full.content == "foo bar"
+
+    # v1
+    llm = GenericFakeChatModel(messages=iter(messages), output_version="v1")
+    full_v1: Optional[BaseMessageChunk] = None
+    for chunk in llm.stream("hello"):
+        assert isinstance(chunk, AIMessageChunk)
+        assert isinstance(chunk.content, list)
+        assert len(chunk.content) == 1
+        block = chunk.content[0]
+        assert isinstance(block, dict)
+        assert block["type"] == "text"
+        assert block["text"]
+        full_v1 = chunk if full_v1 is None else full_v1 + chunk
+    assert isinstance(full_v1, AIMessageChunk)
+    assert full_v1.response_metadata["output_version"] == "v1"
+
+    # v1 from env var
+    monkeypatch.setenv("LC_OUTPUT_VERSION", "v1")
+    llm = GenericFakeChatModel(messages=iter(messages))
+    full_env = None
+    for chunk in llm.stream("hello"):
+        assert isinstance(chunk, AIMessageChunk)
+        assert isinstance(chunk.content, list)
+        assert len(chunk.content) == 1
+        block = chunk.content[0]
+        assert isinstance(block, dict)
+        assert block["type"] == "text"
+        assert block["text"]
+        full_env = chunk if full_env is None else full_env + chunk
+    assert isinstance(full_env, AIMessageChunk)
+    assert full_env.response_metadata["output_version"] == "v1"
+
+
+async def test_output_version_astream(monkeypatch: Any) -> None:
+    messages = [AIMessage("foo bar")]
+
+    # v0
+    llm = GenericFakeChatModel(messages=iter(messages))
+    full = None
+    async for chunk in llm.astream("hello"):
+        assert isinstance(chunk, AIMessageChunk)
+        assert isinstance(chunk.content, str)
+        assert chunk.content
+        full = chunk if full is None else full + chunk
+    assert isinstance(full, AIMessageChunk)
+    assert full.content == "foo bar"
+
+    # v1
+    llm = GenericFakeChatModel(messages=iter(messages), output_version="v1")
+    full_v1: Optional[BaseMessageChunk] = None
+    async for chunk in llm.astream("hello"):
+        assert isinstance(chunk, AIMessageChunk)
+        assert isinstance(chunk.content, list)
+        assert len(chunk.content) == 1
+        block = chunk.content[0]
+        assert isinstance(block, dict)
+        assert block["type"] == "text"
+        assert block["text"]
+        full_v1 = chunk if full_v1 is None else full_v1 + chunk
+    assert isinstance(full_v1, AIMessageChunk)
+    assert full_v1.response_metadata["output_version"] == "v1"
+
+    # v1 from env var
+    monkeypatch.setenv("LC_OUTPUT_VERSION", "v1")
+    llm = GenericFakeChatModel(messages=iter(messages))
+    full_env = None
+    async for chunk in llm.astream("hello"):
+        assert isinstance(chunk, AIMessageChunk)
+        assert isinstance(chunk.content, list)
+        assert len(chunk.content) == 1
+        block = chunk.content[0]
+        assert isinstance(block, dict)
+        assert block["type"] == "text"
+        assert block["text"]
+        full_env = chunk if full_env is None else full_env + chunk
+    assert isinstance(full_env, AIMessageChunk)
+    assert full_env.response_metadata["output_version"] == "v1"
