@@ -214,8 +214,8 @@ def test_rate_limit_skips_cache() -> None:
     assert list(cache._cache) == [
         (
             '[{"lc": 1, "type": "constructor", "id": ["langchain", "schema", '
-            '"messages", '
-            '"HumanMessage"], "kwargs": {"content": "foo", "type": "human"}}]',
+            '"messages", "HumanMessage"], "kwargs": {"content": "foo", '
+            '"type": "human"}}]',
             "[('_type', 'generic-fake-chat-model'), ('stop', None)]",
         )
     ]
@@ -241,7 +241,8 @@ def test_serialization_with_rate_limiter() -> None:
     assert InMemoryRateLimiter.__name__ not in serialized_model
 
 
-async def test_rate_limit_skips_cache_async() -> None:
+@pytest.mark.parametrize("output_version", ["v0", "v1"])
+async def test_rate_limit_skips_cache_async(output_version: str) -> None:
     """Test that rate limiting does not rate limit cache look ups."""
     cache = InMemoryCache()
     model = GenericFakeChatModel(
@@ -250,6 +251,7 @@ async def test_rate_limit_skips_cache_async() -> None:
             requests_per_second=20, check_every_n_seconds=0.1, max_bucket_size=1
         ),
         cache=cache,
+        output_version=output_version,
     )
 
     tic = time.time()
