@@ -48,25 +48,25 @@ def test_prompty_basic_chain() -> None:
     user_message = msgs[1]
 
     # Check the types of the messages
-    assert (
-        system_message["type"] == "system"
-    ), "The first message should be of type 'system'."
-    assert (
-        user_message["type"] == "human"
-    ), "The second message should be of type 'human'."
+    assert system_message["type"] == "system", (
+        "The first message should be of type 'system'."
+    )
+    assert user_message["type"] == "human", (
+        "The second message should be of type 'human'."
+    )
 
     # Test for existence of fakeFirstName and fakeLastName in the system message
-    assert (
-        "fakeFirstName" in system_message["content"]
-    ), "The string 'fakeFirstName' should be in the system message content."
-    assert (
-        "fakeLastName" in system_message["content"]
-    ), "The string 'fakeLastName' should be in the system message content."
+    assert "fakeFirstName" in system_message["content"], (
+        "The string 'fakeFirstName' should be in the system message content."
+    )
+    assert "fakeLastName" in system_message["content"], (
+        "The string 'fakeLastName' should be in the system message content."
+    )
 
     # Test for existence of fakeQuestion in the user message
-    assert (
-        "fakeQuestion" in user_message["content"]
-    ), "The string 'fakeQuestion' should be in the user message content."
+    assert "fakeQuestion" in user_message["content"], (
+        "The string 'fakeQuestion' should be in the user message content."
+    )
 
 
 def test_prompty_used_in_agent() -> None:
